+++ gui/src/app/pages/NwbPage/NwbMainView/SelectedNeurodataItemsWidget.tsx
@@ -34,63 +34,10 @@ const SelectedNeurodataItemsWidget: FunctionComponent<Props> = () => {
       ) : (
         <span>No views selected</span>
       )}
-      <SpecialPSTHButton
-        selectedItemViews={selectedItemViews}
-        openTab={openTab}
-      />
       <SpecialTimeAlignedSeriesButton
         selectedItemViews={selectedItemViews}
         openTab={openTab}
       />
-      <SpecialEphysAndUnitsButton
-        selectedItemViews={selectedItemViews}
-        openTab={openTab}
-      />
-    </div>
-  );
-};
-
-type SpecialPSTHButtonProps = {
-  selectedItemViews: string[];
-  openTab: (path: string) => void;
-};
-
-const SpecialPSTHButton: FunctionComponent<SpecialPSTHButtonProps> = ({
-  selectedItemViews,
-  openTab,
-}) => {
-  const { unitsSelectedItemPath, timeIntervalsSelectedItemPath } =
-    useMemo(() => {
-      if (selectedItemViews.length !== 2)
-        return {
-          unitsSelectedItem: undefined,
-          timeIntervalsSelectedItem: undefined,
-        };
-      const unitsSelectedItemPath = selectedItemViews
-        .find((a) => a.endsWith("|Units"))
-        ?.split("|")[0]
-        ?.split(":")[1];
-      const timeIntervalsSelectedItemPath = selectedItemViews
-        .find((a) => a.endsWith("|TimeIntervals"))
-        ?.split("|")[0]
-        ?.split(":")[1];
-      return { unitsSelectedItemPath, timeIntervalsSelectedItemPath };
-    }, [selectedItemViews]);
-  const enabled = unitsSelectedItemPath && timeIntervalsSelectedItemPath;
-  if (!enabled) return <span />;
-  return (
-    <div>
-      <Hyperlink
-        onClick={() => {
-          if (!unitsSelectedItemPath) return;
-          if (!timeIntervalsSelectedItemPath) return;
-          openTab(
-            `view:PSTH|${timeIntervalsSelectedItemPath}^${unitsSelectedItemPath}`,
-          );
-        }}
-      >
-        Open PSTH for selected items
-      </Hyperlink>
     </div>
   );
 };
@@ -142,49 +89,6 @@ const SpecialTimeAlignedSeriesButton: FunctionComponent<
         }}
       >
         Open TimeAlignedSeries for selected items
-      </Hyperlink>
-    </div>
-  );
-};
-
-type SpecialEphysAndUnitsButtonProps = {
-  selectedItemViews: string[];
-  openTab: (path: string) => void;
-};
-
-const SpecialEphysAndUnitsButton: FunctionComponent<
-  SpecialEphysAndUnitsButtonProps
-> = ({ selectedItemViews, openTab }) => {
-  const { unitsSelectedItemPath, electricalSeriesItemPath } = useMemo(() => {
-    if (selectedItemViews.length !== 2)
-      return {
-        unitsSelectedItem: undefined,
-        electricalSeriesItemPath: undefined,
-      };
-    const unitsSelectedItemPath = selectedItemViews
-      .find((a) => a.endsWith("|Units"))
-      ?.split("|")[0]
-      ?.split(":")[1];
-    const electricalSeriesItemPath = selectedItemViews
-      .find((a) => a.endsWith("|ElectricalSeries"))
-      ?.split("|")[0]
-      ?.split(":")[1];
-    return { unitsSelectedItemPath, electricalSeriesItemPath };
-  }, [selectedItemViews]);
-  const enabled = unitsSelectedItemPath && electricalSeriesItemPath;
-  if (!enabled) return <span />;
-  return (
-    <div>
-      <Hyperlink
-        onClick={() => {
-          if (!unitsSelectedItemPath) return;
-          if (!electricalSeriesItemPath) return;
-          openTab(
-            `view:EphysAndUnits|${electricalSeriesItemPath}^${unitsSelectedItemPath}`,
-          );
-        }}
-      >
-        Open Ephys+Units for selected items
       </Hyperlink>
     </div>
   );